+++ include/utils/WavUtils.hpp
@@ -19,5 +19,5 @@ struct ListTags {
 
 ListTags parseListChunk(const std::vector<char> &value);
 std::string formatMetadataText(const Parser &parser);
-std::string renderWaveformASCII(const std::vector<int16_t> &audioData, int width = 200,
+std::string renderWaveformASCII(const std::vector<float> &audioData, int width = 200,
                                 int height = 30);
